+++ changelog.md
@@ -29,6 +29,8 @@ Codion Change Log
 - CalendarPanel, add/removeListener methods replaced with methods that return the underlying event observer.
 - ColumnConditionPanel, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredTable, add/removeListener methods replaced with methods that return the underlying event observer.
+### is.codion.swing.framework.ui
+- EntityPanel, add/removeListener methods replaced with methods that return the underlying event observer.
 
 ## 0.17.34
 ### is.codion.common.core